requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1

Alembic==1.13.2
psycopg2-binary==2.9.9
Pillow==10.4.0
Flask-Login==0.6.3
Jinja2==3.1.4
SQLAlchemy==2.0.35
gunicorn==23.0.0
Werkzeug==3.0.4
itsdangerous==2.2.0


